Adds LifecycleOwnerWrapper for use-live-data
Summary:
This is the rework of D54364594. One of reasons we need to add a wrap is because we added default lifecycle owner to BaseMountingView(D56757176) and would like to use it for useLiveData, but useLiveData may be called before the view is attached. Adding a default wrapper could solve this issue.

In this case, we need to set LifecycleOwnerWrapper as a TreeProp to all ComponentTree rather than only root ComponentTree

Reviewed By: adityasharat

Differential Revision: D57863409

fbshipit-source-id: 1ba68f8508dca325bbc0ca7b3885f864d2740ec8