java.lang.ArrayIndexOutOfBoundsException:com.facebook.rendercore.transitions.TransitionUtils.addTransitions
Summary:
During a recent refactor, we migrated from having separate transitions and transitions with dependency list in the ReductionState. This refactor introduced a change that ended up being the source of this bug.

Previously, we passed a copy of the transitions list from ResolveResult.Outputs into ReductionState. However, with the change, we were instead passing the same instance: https://fburl.com/code/3klbnqje
Now, as `ResolveResult` may be shared across multiple layouts and in different threads, the side effect of that change is that we were mutating the same transition lists from different threads, leading to the unfortunate crash

With this knowledge, the fix is as simple as ensuring that we pass a copy of the transition data to ReductionState instead.

In addition, I also made a small performance adjustment in MutableTransitionData.add so that it preallocates a right size for the list before calling addAll. Previously, we simply created the list without pre-sizing and called add(item) individually.

Reviewed By: adityasharat

Differential Revision: D58176664

fbshipit-source-id: c6fe52d103fef3ecad273a804a41d77ce5d68e6a