Annotate Hook annotation itself as Unconditional
Summary:
Here, I'm also adding the Unconditional annotation to the Hook annotation itself.
Since Hooks by definition are unconditional, we also want them to
participate in the validation of other Unconditional. For example, it should be possible to call an unconditional from within a hook.
However, we should explicitly avoid Unconditional lint checks for Hook calls themselves since that would result in duplicate signals

Perhaps in future, we can merge the common lint checks from Hook into Unconditional, then we'd be able to remove the explicit Hook logic in the unconditional misuse detector (coming in the following diff)

Reviewed By: fbcbl

Differential Revision: D58191563

fbshipit-source-id: 5650a09626cedf14a4f41f9266ed51c00f033a22