Add pull-to-refresh injection point to Litho
Summary:
This diff along with D58309128 (next one) adds a custom swipe-refresh indictor to FB Litho-based surfaces.
This  is a temporary solution needed to unblock a pull-to-refresh experiment.
After the experiment, this extension point will be replaced with a permanent extendable swipe-refresh layout and a stable Litho API.
This approach is discussed and suggested here: https://fb.workplace.com/groups/litho.support/permalink/3788068054847868/
 https://pxl.cl/52Md9

Reviewed By: pentiumao

Differential Revision: D57927742

fbshipit-source-id: 876d2f90be1e2b7eccc16b677d5f2cd0fe46323e