add config for cloning state list animators in LithoViewAttributesExtension
Summary:
This is another attempt to address the SLA (state list animator) JNI crashes we keep running into iterations of the nested tree experiment. The current hypothesis is that the SLA is shared across views such that one view attempts to update its SLA drawable state while another has unmounted causing the underlying animators to lose their view references.

To verify this hypothesis, we clone any SLAs we're given during application of the `LithoViewAttributesExtension` to ensure that it has a fresh state.

Reviewed By: adityasharat

Differential Revision: D58471256

fbshipit-source-id: fb44f01a4858958a20eb79d7541ca8801ab05814