Remote zoomable functionality
Summary:
What this diff is doing:
- removing everything not needed for the "remote zoom" functionality
- adding some stuff that we need in order to have this top layer with zoom

You pinch to zoom and as long as you are touching the screen the image will be in the top layer with dark background scaling as you go. The moment you release the touch, it animates back to the initial position and it makes the full screen view invisible and the image is back to the component.

Reviewed By: adityasharat

Differential Revision: D58192753

fbshipit-source-id: e5fb5722e68f8d097d0d4763f7a9fe08204a1e5f