Add an option to skip selection when replacing text in TextInput
Summary:
Originally in D29283872 it was [suggested](https://www.internalfb.com/diff/D29283872?dst_version_fbid=519923559226899&transaction_fbid=1680593658800870) to update selection after replacing the text. It makes sense, when text is replaced outside of the input, but auto suggestions are doing it based on the input state, which ends up in a weird situation, when a user changes selection, the suggestion is removed, which means `replaceText` call that will update selection to the last symbol, which will trigger a new suggestion.

Since this feature is already used in several places, it's unclear how full removal of selection will affect all the callsites. Thus, I added an extra field to `ReplaceTextEvent`, so we can skip selection logic only for auto complete plugin.

Reviewed By: adityasharat

Differential Revision: D58726919

fbshipit-source-id: 96db0ef820175affab5a69dcb9afada0bdecac48