Fix bug that cause by double measure pattern.
Summary:
We found a bug in a GenAi integration where we had a pattern of calling measure twice (https://fburl.com/code/87g7k78i):
- Step 1: We measure the children which allows allows the components to take as much size as they needed (and adjust to any changes).
- Step 2:  Measure a second time with an AT_MOST constraint (because of the 75% limit for the bottom sheet height)

I'm terrible at explaining this concept above, so I asked chat gpt to see if it is more clear for the reader:

> The double measure pattern in Android ensures that views are first measured according to their intrinsic content and then adjusted to fit within the constraints of their parent layout. This two-step process helps maintain a flexible yet constrained UI, where views can dynamically adapt to different screen sizes and orientations while ensuring a coherent layout.

This was due to the fact that we were having an optimization that was preventing the content to be synchronously measured in the `measure` call when the root had actually changed (https://fburl.com/code/xbpl9ezj).

```
 if (hasExactSameSpecs || hasSameRootAndEquivalentSpecs) {
          measureOutput[0] = mMainThreadLayoutState.getWidth();
          measureOutput[1] = mMainThreadLayoutState.getHeight();
          needsSyncLayout = false;
        } else {
          needsSyncLayout = true;
 ...
```

In this case we were just checking that `hasExactSameSpecs` were the same, and therefore we were starting an asynchronous resolve. Right after, we would try to constraint, and since the size specs were effectively different we would apply a sync resolve, and just make that final (effectively not verifying if the children had grown in step 1).

## Historical Context

After looking into the source of the code that I am removing I have found that this was pre Split Resolve and Layout and it was introduced [here](https://www.internalfb.com/diff/D7013159?whitespace=SHOW_ALL).

Reviewed By: adityasharat

Differential Revision: D58786257

fbshipit-source-id: 1587565b81b22b370044b7de66ad6a0a41da3e66