Make sure that the PTR is invoked correctly.
Summary:
There is a bug on the *Primitive Recycler* path where the onRefresh callbacks were not correctly invoked and therefore the refresh was not invoked and the spinner was never dismissed.

I couldn't exactly understand why the `static` `Recycler.onRefresh` method was not running when invoked directly, but it worked to delegate the common code into another method and call that one from the Primitive Recycler lambda.

Reviewed By: zielinskimz

Differential Revision: D58813246

fbshipit-source-id: 8f72430044479280a321d2394718d8aa9309416c