Add MountItemPoolsReleaseValidator
Summary:
# Context

I had to investigate potential sources of memory leaks for the Instagram Feed implementation. As I couldn't reproduce any specific memory leak, I have built this  `MountItemPoolsReleaseValidator` which inspects views before they are released into the Pool, and looks for usual sources of leaks: uncleaned click/touch/long click listeners and tags.

This will work for any complex primitives/mount specs that may also forget to clean these view properties from nested views of the Primitive.

# What it doesn't cover

This won't detect if this is any custom view, with particular custom listeners. Covering all possible cases in custom views is just "impossible", so this is just a tool to help with the basic cases.

# How to use

Just go to `MountItemsPools` and initialize it as you wish (you can also add regex to exclude some patterns that you have cleared out as "safe").

An example is:

```
  private val mountItemPoolsReleaseValidator: MountItemPoolsReleaseValidator? =
      MountItemPoolsReleaseValidator(
          excludedPatterns =
              setOf(
                  "IgProgressImageView@id/row_feed_photo_imageview->TextView".toRegex(),
                  "com\\.instagram\\.common\\.bloks\\.components\\.BloksRenderComponent".toRegex(),
                  "^com\\.instagram\\.feed\\.ui\\.rows\\.mediaview\\.MediaPreviewPrimitiveComponent(.*)1->IgFrameLayout$"
                      .toRegex()))

```

One example of a "safe" example, was a custom view in IG that set an internal click listener on init, but never cleaned it. That was a safe listener because it wasn't capturing anything particular that could cause a leak.

## Where this has worked

This has helped me to identify already a couple of sources potential sources of leaks:
- D58944182
- D58922987

Reviewed By: jettbow

Differential Revision: D58945023

fbshipit-source-id: e2058053248d0ca9b4fdcdd9464310e567433677