Automatically clear common view listeners
Summary: This diff sets up an experiment where we'll be testing automatic clearing of common view listeners when the view is being returned to the pool. It won't cover custom views and their listeners, but it'll tell us if there are any potential memory wins we can gain from this.

Reviewed By: fbcbl

Differential Revision: D58995527

fbshipit-source-id: 098e4a45c0de5a8076ceef0f1bbac4f25b62f75d