Attempt to fix content description reannouncement
Summary:
This diff attempts to fix an issue with content description reannounement non-working in some cases.
It's still unclear what is causing the issue but we're making this change to unblock Design Systems team.

It turns out that sending CONTENT_CHANGE_TYPE_CONTENT_DESCRIPTION accessibility event after updating content description fixes the issue but it may potentially break something else so we're adding it behind MC.

Reviewed By: apowolny

Differential Revision: D59054963

fbshipit-source-id: 8a88b42c938de8702f8909240dc4239840cad67f