Put back and gate performSameSize checks
Summary:
# Context

In D58786257 , we have fixed a bug we have identified on a scenario where a double-measure approach is taken, and which was leading to incorrect measurements on the LithoView content.

We are now seeing an increase (S430484) of ANRs on Litho for FB4A (v471), so there is an hypothesis that this fix is actually leading to more synchronous layouts, forcing more work to happen in the Main Thread (namely calculating change sets).

Therefore, we are restoring the original code and adding gating so that we can make the GenAi surfaces to have the fix working for them, while we still verify if this is indeed the cause of increased ANRs and on how we could mitigate it if so.

Reviewed By: adityasharat

Differential Revision: D59290188

fbshipit-source-id: 6d1b31cf65178df665fdcae9f3cac997e9f4550f