Add PoolingPolicy abstraction
Summary:
In this diff we add an abstraction that allows fine-grained control around how content from a `ContentAllocator` can be used with pooling:
- If it can be released into a pool
- If it can be acquired from a pool

This will allow us to enable the following use case:
- enabling background pre-allocation of `ComponentHost`, while still preventing them to be released into the pool due to the known problems we have in recycling them.

This opportunity showed up in the scope of the "disabling BG/FG outputs" which is seeing regressions, likely due to the increased number of hosts that have to be created. Our hypothesis is that doing this could alleviate this regression without having bugs of host recycling.

Reviewed By: adityasharat

Differential Revision: D59223605

fbshipit-source-id: 68abcb4e0db2bad128f2aaa123138eb01e27aa72