Make pools synchronized
Summary:
We are seeing a crash on rendercore which indicates that some `null` content is being retrieved from the `Pool` (https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:core/core/src/main/java/androidx/core/util/Pools.kt;l=83?q=Pools.kt) . However, our code is in Kotlin and doesn't support `null` content to be released into the pool.

This issue started after the upgrade of the `androidx` libraries where they have changed the `SimplePool` to not support `null` items (D58951507)

We believe the issue is because of race conditions on `release` method below:

```
1+ override fun release(instance: T): Boolean {
2+  check(!isInPool(instance)) { "Already in the pool!" }
3+  if (poolSize < pool.size) {
4+     pool[poolSize] = instance
5+     poolSize++
6+     return true
7+  }
8+  return false
9+}
```

(Assuming an empty pool at the start)

- **Thread A (Main Thread)** calls `release(instanceA)` and `poolSize=0`, therefore it will invoke `pool[0] = instanceA`
- **Thread B (BG Thread)** calls `release(instanceB)` and sees that `poolSize=0` still (the `Thread A` has not incremented it yet), and therefore calls `pool[0]= instanceB`
- Both Threads eventually increment the `poolSize`, so `poolSize=2`, but there is only element in the pool: `[instanceB, null]`
- **Thread A** calls acquire later on, and tries to retrieve `pool[poolSize-1]`, which is `null`.

This is likely leading to a change in the `lastPooledIndex` in the `SimplePool` code, which is then causing us to access an incorrect and null value.

Reviewed By: pentiumao

Differential Revision: D59397596

fbshipit-source-id: d74837e87c5bf9f87c6fc0d49be173b28103b3de