add pass StaggeredGridLayoutInfoFactory to LazyStaggeredGrid
Summary:
## Context
zhiweicheng noticed that the crossAxis padding between the column in staggered grid is wrong.

We were using itemDecoration(https://fburl.com/code/ac0nza2z) to apply the padding on the edge of the view based on its spanIndex, which is supposed to be the common practice to set padding between views in recyclerView. When recyclerView applies itemDecoration, it should also re-measure the width and height of the child view accordingly.

However, this re-measure in crossAxis(which is width in our case) is not working in LithoView and the re-measured size is always overrode at https://fburl.com/code/dhsledjr.

This means that we should set up the width of the child from the beginning instead of relying on re-measurement when recyclerView applies itemDecoration.
## This diff
- Pass StaggeredGridlayoutInfoFactory to LazyStaggeredGrid, so that we can change the width of child view.

## The bug in spacing
https://pxl.cl/5bSN3

Differential Revision: D59413730

fbshipit-source-id: d6654ad2e564e4105c0b04aefe4e58302f74dcd8