Allow multiple zoom components on one screen
Summary:
This diff seems huge but majority of it is just changes from previous zoom controller to the new one so dont be afraid!

In this diff we are changing the approach and we are passing touch detector instead of a controller. Thanks to that we can have multiple zoomable components that are using one touch detector (instagram use case).
We have a listeners list and all of the views that are currently on the screen will receive the touch, then they will decide if this touch is interesting to them (if the touch happened in their bounds, then yes)

Reviewed By: zielinskimz

Differential Revision: D59371856

fbshipit-source-id: 12c8dd16e9c979ff4ae96f4cad49be60022ff12f