Wrap exceptions happening during mount/unmount with metadata about RU and the binder
Summary:
# Context
Whenever there is an exception happening during mount/unmount on *RenderCore*, we have no good mechanisms to understand the source of it (e.g, which particular render unit it's causing it).

Therefore, I'm adding an handling on the main mount/unmount callsites so that if an exception happens, we catch it and rethrow it with information on the *RenderUnit* description and optionally about the binder (if we understand if it's coming from a particular binder).

Reviewed By: zielinskimz

Differential Revision: D59677351

fbshipit-source-id: b061bba24b6d83efbbeed00b44e5b4dc788bd2c3