Add proper link to the cheatsheet for injection docs
Summary:
Just changing the link to the proper one

This diff has been automatically generated by the inpage editor.

NOTE: If you want to update this diff, go via the preview link inside the static docs section below.
Ensure you are editing the same page that was used to create this diff.

Reviewed By: kingsleyadio

Differential Revision: D59753596

fbshipit-source-id: f5f4eece345d21511473748cfbe7b18ca5a11d7d