Implement back animation to the initial place
Summary:
Previously if we would drag the image we would always animate to the middle and then scale. now we scale and animate back at once.
In this diff i switched from two springs that were responsible for X and Y position, and used just one sprign for back animation together with calculations of the proper position.

Reviewed By: zielinskimz

Differential Revision: D59925807

fbshipit-source-id: bd90eb3b3b711b058c44780639d5e35767cc5e69