Use LifecycleOwner to clear MountItemsPool
Summary:
MountItemsPool uses `Context` as a map key to store the content Pools. The map is static which means that once we store something inside, it will stay there forever unless we remove it. In case where the Context is an Activity, we're registering ActivityLifecycleCallbacks in order to listen to all activity create and destroy events and we can use these callbacks to clear the pool when the activity is destroyed in order to avoid memory leaks.
But there are cases when the Context is not an activity but for example a Service. And in such case we would never remove the entry from static map and this leads to memory leaks as reported in this on-call request: https://fb.workplace.com/groups/litho.support/posts/3814655782189095/

There is no LifecycleCallbacks API for the Service but there is the [LifecycleService](https://developer.android.com/reference/androidx/lifecycle/LifecycleService) class in AndroidX Lifecycle library. So when the client apps uses this for their Service implementation, the Service becomes a lifecycleOwner and we can use that to clear the pool when the service is destroyed.

In case the client code can't extend from LifecycleService, they can use [ServiceLifecycleDispatcher](https://developer.android.com/reference/androidx/lifecycle/ServiceLifecycleDispatcher) directly in their Service to dispatch the lifecycle events similar to how `LifecycleService` does it [internally](https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:lifecycle/lifecycle-service/src/main/java/androidx/lifecycle/LifecycleService.kt;l=26).

This diff adds the logic that should handle cases where the Context is a LifecycleOwner.

Reviewed By: jettbow

Differential Revision: D59673321

fbshipit-source-id: 93baf548f65afe21d7b1bf00d352a0be4e8ee921