Move binder.unmount call to the end.
Summary:
In order to have some consistency on approach, we are moving the binder.unmount call to the end of the unmount block. This will also allow us to extract the binding code into helper methods (as in the next diff of the stack) so that we can share better during the experimentation with other strategies for the ExperimentalRecycler.

After some verification, it seems to me that there is no direct relationship of order, and that we can invoke the `binder.unmount` call later.

Reviewed By: zielinskimz

Differential Revision: D60131048

fbshipit-source-id: 0ea7ee9b5cf6ecf7c67b69dbc9bed70ca2f55352