Add ability to log non fatals regarding improper mount release
Summary:
I'm adding a new non-fatal with the goal of identifying improper mount content releases into the Pool on IG. Hopefully this will be able to signal some situations that we are not being able to repro and then fix them, so as to avoid leaks.

This will be only enabled for a small percentage of users, to avoid noise.

Differential Revision: D60234842

fbshipit-source-id: 5d63809ede626b7cef49fccfb4f353b4178cee6f