Use after pooling fields state to identify changes on fields state before release.
Summary:
In order for a more accurate verification of the improper release of items, we are adding extra checks.

We are now recording the state of the fields right when the item is created/pooled. This state is then compared to the one at the moment of the item release back into the pool. This allow us to ignore situations where there is a field/listener that is already set from the initial moment (as part of custom view for example).

Reviewed By: adityasharat

Differential Revision: D60234841

fbshipit-source-id: 7f4cd8a7d5a8d2f09365e8b90663b503c9fedf66