Refactor Crashing Mount Pool Helper to Support Other Components
Summary:
## This Diff

Support other components to be supplied to the crashing mount spec helper.

## More Context

In the next diff, I'm going to remove support for `InjectProp` as an optional parameter in delegation methods.  In doing so, we can no longer depend on it for crashing mount specs.  Therefore, instead of making these inject props, I'm going to keep the existing tests by creating new mount specs that crash in the designated functions.

Therefore, as a first step, this splits out the refactor to support changing the component.

Reviewed By: apowolny

Differential Revision: D60273237

fbshipit-source-id: 9ecfd41a4bf96dbca8c24c64639ddb5c6c6a3e16