Add privateImeOptions
Summary: We would like to disable the gif/sticker buttons in device keyboard(see T159878027. Gif and sticker are not supported for story reply, when user click on any gif or sticker from device keyboard, it will just do nothing and cause confusion for the user.), so we need to set privateImeOptions for the editText in TextInputSpec, this diff exposes it as a prop for the component.

Differential Revision: D60190252

fbshipit-source-id: bba2f9717bc599d88d05b9fc525b47c1c60779ca