Fix mount item pools lifecycle owner threading
Summary:
The content can be acquired from the pool on background thread and when that happens Litho ensures that an observer is added to the LifecycleOwner in order to be able to clear the pool when the owner is destroyed.
The issue is that lifecycleOwner.addObserver has to be called on main thread, so this diff ensures that observer is always added on the main thread.

The lifecycle owner code in mount items pool is behind MC and the experiment hasn't started yet, so no users were affected by this issue so far.

Reviewed By: adityasharat

Differential Revision: D60436110

fbshipit-source-id: 28de13768baa7dbc0980df7d3554aa6539bd6fcb