Avoid false positives in mount pool validation
Summary:
In the internal of Litho we use an indirection technique where we have always `ComponentLongClickListener` or `ComponentTouchListener` and we only nullify the inner `eventHandler` whenever a view is unmounted.

I'm adding support in the Litho IG integration to detect these situations correctly, otherwise we might be reporting false positives.

Reviewed By: pentiumao

Differential Revision: D60814806

fbshipit-source-id: 8b2fa0631cb9349428ad67365411d637b7066e02