Add pagination support for recycler
Summary:
Context:

This diff is to address the long-standing issue with pagination(https://fb.workplace.com/groups/litho.support/permalink/3818210841833589/), where product team has to deal with the case that scrolling to the correct position while inserting new items, which introduced complexity and caused a lot of bugs.

The main idea of the solution is, when having dataset change, more specifically when insertion happens, we could let recycler view to remain to the last visible position, then product team doesn't need to handle the pagination event anymore because we always stay at the position of freshly inserted item. (where loading item was)

The downside of the solution is: it only works for full-screen size surface like Reels, otherwise, user would notice automatic snap behaviors while scrolling.(when insertion happens on the background)

Reviewed By: adityasharat

Differential Revision: D60760993

fbshipit-source-id: cf3e8c2baaff6395cc6d15b5d559da8d25ce5962