Pass context while creating render unit
Summary:
in the past we used to get OOMs when we held cc on the host and drawable component. But we didnt have render units back then, so now it shoudnt happen.

We discovered this as an issue while working on focus order api, where i needed context in order to traverse the tree while looking for a tagged view.

Reviewed By: adityasharat

Differential Revision: D61123872

fbshipit-source-id: 5730b3ccbdc6a73295fa005ec6fc4719688493b7