Attempt to fix null LayoutContext crash
Summary:
While running an experiment which enables Primitive measurement logic fix, we discovered a [crash](https://www.internalfb.com/logview/facebook_android_crashes/ffe987ce1ebc1d746e4acd0524834dfc?trace_tab=0&ds=%7B%22start%22%3A%22-2%20weeks%22%2C%22end%22%3A%22now%22%7D).

We're not able to reproduce it but our hypothesis is that this is what happens:
1. There is Component that calls measure of a component during render (which adds that component's to the measured layout cache)
2. When measured layout is used it is remove from the cache (but not from the delegate cache)
3. After layout the tree is walked and the context is dereferenced

When the next layout happens and the measured layout is found in the cache it crashes when it is measured because the clear cache method didn't remove the layout from the delegate cache.

This diff aims to fix it by removing the layout from the delegate cache.

Reviewed By: adityasharat

Differential Revision: D61582329

fbshipit-source-id: f0c66d0dca1b4ae83593ec5f1e7358f9b606ad1c