Rename ImageSpec to ImageComponentSpec
Summary:
This diff renames `ImageSpec` to `ImageComponentSpec` and adds `ImageSpec` LayoutSpec component that delegates to `ImageComponentSpec`. It also removes an unused `useIntrinsicSize` API.

The goal is to replace `ImageSpec` with a `PrimitiveComponent` implementation. With this setup we won't have to change the callsites and we'll be able to switch from `ImageComponentSpec` to PrimitiveComponent easily.

Reviewed By: adityasharat

Differential Revision: D60749339

fbshipit-source-id: 367ac773724600823b8ffec67a0d5a685c07f765