Detect and log instances of duplicated render unit crashes
Summary:
See https://fb.workplace.com/groups/bloks.bugs.regressions/permalink/8339876572722604/ for context

Adding the ability to annotate the exception with the scope key for the model that is implicated in the exception.

This does not catch all code paths but should help us narrow down the specific case in the post above.

Reviewed By: pasqualeanatriello

Differential Revision: D62125761

fbshipit-source-id: d31264cf860002d2c9899a65de1799d71671ebe1