Revert D62240809: Multisect successfully blamed "D62240809: [Codemod] Add Nullchecker Flag to IG4A Targets" for one test failure
Summary:
This diff reverts D62240809
D62240809: [Codemod] Add Nullchecker Flag to IG4A Targets by jocelynluizzi13 causes the following test failure:

Tests affected:
- [fbsource//fbandroid/java/com/facebook/cameracore/ardelivery/voltron:dep_test - main](https://www.internalfb.com/intern/test/281475090911831/)

Here's the Multisect link:
https://www.internalfb.com/multisect/9703659
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Differential Revision: D62271625

fbshipit-source-id: 5538ab28f869e233b5669dc5d824d9637f1eb205