Uses cheaper check if event handler redesign is enabled
Summary: Uses cheaper check if event handler redesign is enabled. This works because the design only rebinds the event handler when the layout state is promotted to the UI. In the control behaviour the event handlers were rebound on commit; this removed unused event handlers from the controller. On the next remount the event handler would not get rebound because the controller has already released them.

Reviewed By: fbcbl, zielinskimz

Differential Revision: D62376600

fbshipit-source-id: 0012d96c2dccf78c147c2e11992d03e6b396e239