Add compose tracer
Summary:
We are starting to add compose litho component, this diff added the compose trace support, which hope to help the perf checking.

why not use the existing trace library?
the current one is using fbsystrace, which litho sample app doesn't support it.

This diff use the existing trace in the Litho and start it when App init.

Reviewed By: adityasharat

Differential Revision: D62502663

fbshipit-source-id: 2165f869ac77c0e3aab1c70b613fa7c0ca92b7db