Revert D62931034: Multisect successfully blamed "D62931034: [Litho][BE] Remove redundant Android version checks" for one test failure
Summary:
This diff reverts D62931034
D62931034: [Litho][BE] Remove redundant Android version checks by zielinskimz causes the following test failure:

Tests affected:
- [fbsource//fbandroid/instrumentation_tests/com/facebook/fds/generated/product/feed/posttext:posttext_darkmode - feedContentTextMultipleLinesTruncationDarkMode (com.facebook.fds.generated.product.feed.posttext.Fb4aFDSPostTextScreenshotTestDarkMode)](https://www.internalfb.com/intern/test/281475135981882/)

Here's the Multisect link:
https://www.internalfb.com/multisect/10481197
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: zielinskimz

Differential Revision: D63023067

fbshipit-source-id: 0ed255b33d6d7f153981347acd5128e5c493801a