Remove falling test from StateEquivalence
Summary:
In order to fix memory issue we had to remove Component Context from State. We no longer have access to Litho Tree.
I followed up with Adi about safety of this change, and it seems to be safe.

As a followup I will add a lint to detect if we are adding deps to useCached if we do:

```
 val number = useState { 1 }
    val expensiveValue = useCached {
     number.update(some_value)
      expensiveRepeatFunc("hello", number.value)
    }

```
to avoid

Reviewed By: pentiumao

Differential Revision: D63261506

fbshipit-source-id: 826240bbbecfbd8ebaad487f21c8cf32c63ee1c2