Replace set view tags with add view tags
Summary:
Resubmitting my diffs with focus order after they got reverted.
This is a part of the diff that got accepted here; D60849676

In focus order we will be adding view tags.
Currently when we are setting them, we replace the whole array. In this diff we are merging the arrays of view tags and only override the tag that has duplicated key.

Reviewed By: adityasharat

Differential Revision: D63273785

fbshipit-source-id: b218f6e6b0ecaa1486aad19d9ea7fa5a84cec9e1