Fix the bug that state containers get lost for NestedTree
Summary:
# Context
This is to fix a bug that was reported in an oncall request(https://fb.workplace.com/groups/litho.support/permalink/3877583855896287/).

Upon investigation, we noticed that we never tell TreeState to retain state containers for cached NestedTree, which ends up getting lost of state containers and making the first state update has no effect.

Reviewed By: adityasharat, zielinskimz

Differential Revision: D63460548

fbshipit-source-id: 890c518cbb84591001472132a6f854d1e649a3e5