Remove custom content description handling
Summary:
we are trying to fix the contentdescription reannouncement bug.
This custom handling was added back in 2016. Currently all of the functionality would be performed in super.setContentDescription.
https://android.googlesource.com/platform/frameworks/base/+/master/core/java/android/view/View.java#11863

while looking at the method it does exactly the same as we were doing, checking the content of content description, setting it if its different and invalidating the view so it should be safe to delete

Reviewed By: kingsleyadio

Differential Revision: D63534651

fbshipit-source-id: 4c9cc95401c559f02916cc84ee733bc2fdd93c59