Fix the bug where visibility event wasn't triggered for the use case of VisibilityBoundsTransformer
Summary:
This diff is to fix the problem reported by support request(https://fb.workplace.com/groups/litho.support/permalink/3835721820082491/), where fb feed is using bounds transformer to shift the bounds of items in the list and Litho doesn't take account of transformer.

The solution is avoiding early return for the use case of transformer and continue the process of visibility extension.

Reviewed By: adityasharat

Differential Revision: D64041592

fbshipit-source-id: 75374d253cdb7a306f3cf30de7a86157d04b7d47